+++ mod.ts
@@ -5,6 +5,7 @@
  */
 export * from "./federation/mod.ts";
 export * from "./httpsig/mod.ts";
+export * from "./nodeinfo/mod.ts";
 export * from "./runtime/mod.ts";
 export * from "./vocab/mod.ts";
 export { type ResourceDescriptor } from "./webfinger/mod.ts";