+++ nodeinfo/mod.ts
@@ -1,0 +1,7 @@
+/**
+ * This module implements the [NodeInfo](https://nodeinfo.diaspora.software/)
+ * protocol.
+ *
+ * @module
+ */
+export { type NodeInfo } from "./types.ts";